refactor all in one item module
