+++ src/main/java/com/lothrazar/cyclicmagic/component/storagesack/ContainerStorage.java
@@ -25,7 +25,7 @@ package com.lothrazar.cyclicmagic.component.storagesack;
 import com.lothrazar.cyclicmagic.data.Const;
 import com.lothrazar.cyclicmagic.gui.base.ContainerBase;
 import com.lothrazar.cyclicmagic.gui.slot.SlotItemRestrictedInverse;
-import com.lothrazar.cyclicmagic.module.ItemToolsModule;
+import com.lothrazar.cyclicmagic.module.ItemModule;
 import com.lothrazar.cyclicmagic.util.UtilPlayer;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.InventoryPlayer;
@@ -51,7 +51,7 @@ public class ContainerStorage extends ContainerBase {
         x = pad + k * Const.SQ;
         y = pad + l * Const.SQ;
         slot = k + l * cols;
-        this.addSlotToContainer(new SlotItemRestrictedInverse(invoWand, slot, x, y, ItemToolsModule.storage_bag));
+        this.addSlotToContainer(new SlotItemRestrictedInverse(invoWand, slot, x, y, ItemModule.storage_bag));
       }
     }
     int yBase = pad + rows * Const.SQ + 14;
@@ -89,7 +89,7 @@ public class ContainerStorage extends ContainerBase {
   @Override
   public ItemStack transferStackInSlot(EntityPlayer par1EntityPlayer, int index) {
     ItemStack itemstack = ItemStack.EMPTY;
-    Slot slot = (Slot) this.inventorySlots.get(index);
+    Slot slot = this.inventorySlots.get(index);
     if (slot != null && slot.getHasStack()) {
       ItemStack itemstack1 = slot.getStack();
       itemstack = itemstack1.copy();
